backend/requirements.txt
fastapi>=0.08.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
# Kf_predle>=8.3.0  # Remove this invalid line
requests>=1.26.0
numpy>=1.21.0
pydantic>=1.8.0
tenacity>=8.2.2
sentence-transformers>=2.2.2
Jinja2>=3.1.2
oracledb>=2.0.0  # Add this line